Feat: 드디어 와우 숙소 input 값 데이터 받아오기 성공
