+++ src/components/coupon-registration/coupon-applier/common-quantity-coupon-setter/type.ts
@@ -1,0 +1,6 @@
+import { CouponApplierProps } from '../type';
+
+export type CommonQuantityCouponSetterProps = Pick<
+  CouponApplierProps,
+  'allQuantityValue' | 'setAllQuantityValue'
+>;
