+++ src/components/coupon-registration/coupon-applier/index.tsx
@@ -2,20 +2,36 @@ import styled from 'styled-components';
 import { CommonQuantityCouponSetter } from './common-quantity-coupon-setter';
 import { RoomCouponApplier } from './room-coupon-applier';
 import { useCoupon } from '@hooks/coupon/useCoupon';
+import { CouponApplierProps } from './type';
+import { useEffect } from 'react';
 
-export const CouponApplier = () => {
+export const CouponApplier = ({
+  allQuantityValue,
+  setAllQuantityValue,
+  itemQuantityValue,
+  setItemQuantityValue,
+}: CouponApplierProps) => {
   const { couponRoomListData } = useCoupon();
+
+  useEffect(() => {
+    console.log('itemQuantityValue', itemQuantityValue);
+  }, [itemQuantityValue]);
 
   return (
     <Container>
-      <CommonQuantityCouponSetter />
+      <CommonQuantityCouponSetter
+        allQuantityValue={allQuantityValue}
+        setAllQuantityValue={setAllQuantityValue}
+      />
       <StyledRoomCouponApplierWrap>
         {couponRoomListData?.rooms.map((item, index) => (
           <RoomCouponApplier
             key={index}
-            roomId={item.roomId}
             roomName={item.roomName}
             index={index}
+            roomId={item.roomId}
+            itemQuantityValue={itemQuantityValue[index]}
+            setItemQuantityValue={setItemQuantityValue}
           />
         ))}
       </StyledRoomCouponApplierWrap>
