+++ src/components/coupon-registration/coupon-applier/type.ts
@@ -1,12 +1,11 @@
-export type RoomCouponApplierProps = {
-  roomName: string;
-  index: number;
-  roomId: number;
+export type CouponApplierProps = {
+  allQuantityValue: string;
+  setAllQuantityValue: React.Dispatch<React.SetStateAction<string>>;
   itemQuantityValue: {
     roomId: number;
     roomName: string;
     quantity: string;
-  };
+  }[];
   setItemQuantityValue: React.Dispatch<
     React.SetStateAction<
       {
