+++ src/components/coupon-registration/coupon-type/type.ts
@@ -1,3 +1,13 @@
-export type StyledDiscountButtonProps = {
-  className: string | null;
+import {
+  DiscountPriceType,
+  DiscountRateType,
+} from '@/constants/coupon-registration/type';
+
+export type CouponTypeProps = {
+  selectedDiscountType: DiscountPriceType | DiscountRateType;
+  setSelectedDiscountType: React.Dispatch<
+    React.SetStateAction<DiscountPriceType | DiscountRateType>
+  >;
+  discountValue: string;
+  setDiscountValue: React.Dispatch<React.SetStateAction<string>>;
 };
