+++ src/hooks/coupon-registration/useCouponPreviewProvider.tsx
@@ -1,0 +1,4 @@
+export const useCouponPreviewProvider = () => {
+  console.log('!@#');
+  return <div>useCouponPreviewProvider</div>;
+};
