+++ src/pages/coupon-registration/index.tsx
@@ -4,11 +4,28 @@ import { CouponPreview } from '@components/coupon-registration/coupon-preview';
 import { CouponType } from '@components/coupon-registration/coupon-type';
 import { Spacing } from '@components/spacing';
 import { TextBox } from '@components/text-box';
-import { useCouponRoomProvider } from '@hooks/coupon-registration/useCouponRoomProvider';
+import { useState } from 'react';
 import styled from 'styled-components';
+import { DISCOUNT_PRICE_TYPE } from '@/constants/coupon-registration';
+import {
+  DiscountPriceType,
+  DiscountRateType,
+} from '@/constants/coupon-registration/type';
 
 export const CouponRegistration = () => {
-  const { couponQuantitiesByRoom } = useCouponRoomProvider();
+  const [selectedDiscountType, setSelectedDiscountType] = useState<
+    DiscountPriceType | DiscountRateType
+  >(DISCOUNT_PRICE_TYPE);
+  const [discountValue, setDiscountValue] = useState<string>('');
+  const [allQuantityValue, setAllQuantityValue] = useState('0');
+  const [itemQuantityValue, setItemQuantityValue] = useState<
+    {
+      roomId: number;
+      roomName: string;
+      quantity: string;
+    }[]
+  >([]);
+
   return (
     <Container>
       <StyledLeftWrap>
@@ -18,11 +35,21 @@ export const CouponRegistration = () => {
         <Spacing space="8" />
         <StyledCouponCardWrap>
           <CouponCard title="1. 쿠폰 유형 선택">
-            <CouponType />
+            <CouponType
+              selectedDiscountType={selectedDiscountType}
+              setSelectedDiscountType={setSelectedDiscountType}
+              discountValue={discountValue}
+              setDiscountValue={setDiscountValue}
+            />
           </CouponCard>
           <Spacing space="32" />
           <CouponCard title="2. 적용 객실 선택">
-            <CouponApplier />
+            <CouponApplier
+              allQuantityValue={allQuantityValue}
+              setAllQuantityValue={setAllQuantityValue}
+              itemQuantityValue={itemQuantityValue}
+              setItemQuantityValue={setItemQuantityValue}
+            />
           </CouponCard>
         </StyledCouponCardWrap>
       </StyledLeftWrap>
